feat: Minimum Viable Product (#1)
This MVP includes a very basic subset of what is generally planned.

Before merge, this should be able to
- [x] Plugins
  - [x] Load Plugins
- [x] ~~Generate JS Bundle on the Fly from source (rebuilding on change
is out of scope for now)~~ (too insecure, risks arbitrary code execution
due to needing to invoke node / bun
- [x] Expose Plugin Bundle (Code, Codemap, Static Assets) via local
server
  - [x] Display current state of a Plugin during Building
  - [x] Display Plugins as part of the main UI
  - [x] Emit Events to the plugins
- [x] Settings
  - [x] Let each plugin expose a settings UI
  - [x] Enable/Disable Plugins in the settings
As part of the MVP, no bundled / internal plugins are provided. This
will be added later