feat: implement core plugin including basic i18n support (#14)
this includes a UI Component that lets you open the Station, System or
your Ship,
The settings pane now supports toggling the preferred website to open a station/system/ship.
The settings pane also supports selecting a Locale, which gets currently written, but not applied by the Plugin or EDPF itself.
Also, this does not include EDDN integration yet